Fix ROUND() type casting for similarity scores
Fixed PostgreSQL error in duplicate detection query:
- similarity() returns double precision, ROUND() needs numeric
- Added ::numeric cast before ROUND()
- Fixes: function round(double precision, integer) does not exist

Changed:
  ROUND(model_similarity * 100, 1)
To:
  ROUND((model_similarity * 100)::numeric, 1)

+++ queries/component_cleanup_analysis.sql
@@ -111,8 +111,8 @@ SELECT
   comp2_brand,
   comp2_quote_usage + comp2_po_usage AS comp2_total_usage,
 
-  ROUND(model_similarity * 100, 1) AS model_match_pct,
-  ROUND(desc_similarity * 100, 1) AS desc_match_pct,
+  ROUND((model_similarity * 100)::numeric, 1) AS model_match_pct,
+  ROUND((desc_similarity * 100)::numeric, 1) AS desc_match_pct,
 
   -- Recommendation
   CASE
